Fix compiler error
A static const class member in a cout statement results in an undefined symbol
when linking an executable using the class.
The solution is to do a static cast of the variable in the cout statement.

+++ MQ/monitor/CbmDeviceMonitorReqT0.cxx
@@ -76,7 +76,7 @@ try {
 
   if ("" == fsTsBlockName) {
     //
-    LOG(info) << "Requesting TS using the SysId: 0x" << std::hex << kusSysId << std::dec;
+    LOG(info) << "Requesting TS using the SysId: 0x" << std::hex << static_cast<int>(kusSysId) << std::dec;
   }
   else {
     //
